Change default port to 5017
This port number was selected on random from a port range 5016-5019.

More info: https://github.com/shakacode/cypress-on-rails/issues/20

According to [this list](https://www.iana.org/assignments/service-names-port-numbers/service-names-port-numbers.txt) there are a few ranges of ports that are not assigned, the closest to the current setting is 5016-5019, and this [wikipedia page](https://en.wikipedia.org/wiki/List_of_TCP_and_UDP_port_numbers) seems to confirm it.