Mention the correct file logger in the warning
The clean_db file no seems to have been renamed to just clean. I've changed to to the correct file name and because clean is a common word I've added more of the path to make it obvious it is referring to a file and to direct the user reading the warning to the right place.