Update the Rails test server port
Running the next command `yarn cypress open --project ./spec` starts up Cypress which seems to look for a server responding on port `5002` by default rather than the previously documented `bin/rails server -e test -p 5017`. This fix updates the documentation to get the Rails test server running on the port that Cypress expects it to be responding on, `5002`.